# backend/src/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "SnookerPocket API with Swagger",
    "description": "api description",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/speeldagen/": {
      "get": {
        "tags": [
          "Speeldag"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": true
      },
      "post": {
        "tags": [
          "Speeldag"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "schiftingsvraag": {
                  "example": "any"
                },
                "schiftingsantwoord": {
                  "example": "any"
                },
                "wedstrijden": {
                  "example": "any"
                },
                "speeldagVotes": {
                  "example": "any"
                },
                "klassement": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "deprecated": true
      }
    },
...
    "/api/wedstrijden/": {
      "get": {
        "tags": [
          "Wedstrijd"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": true
      },
      "post": {
        "tags": [
          "Wedstrijd"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "datum": {
                  "example": "any"
                },
                "resultaat": {
                  "example": "any"
                },
                "thuis": {
                  "example": "any"
                },
                "uit": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "deprecated": true
      }
    },
...
    "/api/speeldagVotes/": {
      "get": {
        "tags": [
          "SpeeldagVote"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": true
      }
    },
...
    "/api/wedstrijdVotes/": {
      "get": {
        "tags": [
          "WedstrijdVote"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": true
      },
      "post": {
        "tags": [
          "WedstrijdVote"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "vote": {
                  "example": "any"
                },
                "wedstrijd": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "deprecated": true
      }
    },
...
  },
...
}